fixed registration problem

+++ playlist_generator/src/main/java/track_ninja/playlist_generator/models/Authority.java
@@ -30,7 +30,7 @@ public class Authority implements GrantedAuthority {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = AUTHORITY_ID)
-    private int authorityId;
+    private long authorityId;
 
     @Column(name = ROLE_TYPE)
     @NotNull
